requirements.txt
numpy==1.17.0
opencv_python==4.1.0.25
Flask_Cors==3.0.8
Werkzeug==0.15.5
Flask==1.1.1
eventlet==0.25.0
gevent_socketio==0.3.6
